refactor: split i18n dictionary
